requirements.txt
streamlit==1.35.0
requests==2.31.0
pillow==10.3.0
numpy==1.24.4
pandas==1.5.3
python-dotenv==1.0.1